Introduce ProcessingQueue
The main goal of ProcessingQueue is decoupling processing promises (promises which resolves when chunk is ready)
from read promises (promise which is returned when external source reads from IFCA). This way read promise can be
resolved by any processing promise which allows skipping chunks (e.g. filtering). The end result is that input and output
is bound together not on write but on read giving greater flexibility.