Rework transformations chaining
This commit reworks how transformations are handled. The logic is practically the same but the code itself
had been refactored to allow filtering "DroppedChunks" on various stages of transforming.

I have rewritten ".makeSynchronousChain()" and ".makeTransformChain()" (which had been renamed to ".makeAsynchronousChain()"
for consistency) to use "for..of" loops instead of "Array.reduce()" to allow breaking from the loop early if "DroppedChunk"
occurs. This way we don't have to use a workaround to skip reduce iterations for dropped chunks.

The 2nd change is rewritting ".makeStrictTransformChain()" - it now groups sync/async functions and calls them only whenever
next transform is of different type. It uses ".makeSynchronousChain()" and ".makeAsynchronousChain()" internally so there
is no duplicated logic (as it was before for handling async transforms there).