Support async reducers in 'DataStream.reduce'
I have added support for async reducers. This required some iffology
related to detetcing if reducer is sync or async function. In functions
which are called for every chunk we don't want to use 'await' if it's
not needed (due to performance) and so this requires branching code
depending on callback sync/async nature.

I have also duplicated '.getReader()' due to the same reason. It's kind
of longish now and I'm not really proud of it, but additional iffology
inside this method would make it hard to digest.