Fix static and runtime return type for StringSplit for inherited methods
This commit is trying to solve the issue of "polymorphic generic this
as a return type".

For more in-depth description see:
https://github.com/scramjetorg/scramjet-dev/pull/28.

This issue is related to a fact that in `DataStream<T>` class we have
multiple methods (transfroms) which returns `DataStream<U>` (so the
instance of the same class, but with different type parameter). This
works fine for `DataStream` itself but causes problems for derived
classes as a return type is not dynamic and inhertied methods called
on an instance of derived class will always return the instance
of a parent class.

First change was to introduce `.create()` method which is used inside
tramsforms directly instead of calling `new DataStream...`. This allows
derived classes to override it and have control over how its instances
are created. With `.create()` in use, new instance based on call context
(current this) is created. This results with a correct return type
during runtime.

The second change was to override transforms signatures in the derived
class so correct return type could be declared. Those overriden
transforms call `super.transform(...)` internally so the logic stays
exactly the same. This solves the issue of TS not correctly recognizing
return type (as without it, method signatures declare base class
as return type).

Above works well for generic derived classes. However, for non-generic
derived classes (like `StringStream` in our case), overridden transform
signatures (and also `.create()` method) needs to be non-generic as well
to work with this approach. And so the third change was to overload
transforms in a base class to have both non-generic and generic versions.