Introduce write/read methods and stream type transforming
Utilize 'IFCAChain' so  interemediatte streams with shared IFCA
can be created easier. It also allows writing/reading on any
intermediatte stream to use correct IFCA instance - first in the chain
for writing, last in the chain for reading.

Apart from that, new stream instance is always created for each
transform which allows to manage readability/writiability (and similar)
separately.